+++ src/display_player_tree.c
@@ -1,6 +1,6 @@
 #include "../main.h"
 
-void display_player_tree(Player* root, int space) {
+void display_player_tree(Player* root, size_t space) {
     if (root == NULL)
         return;
 
@@ -9,7 +9,7 @@ void display_player_tree(Player* root, int space) {
     display_player_tree(root->right, space);
 
     printf("\n");
-    for (int i = 10; i < space; i++)
+    for (size_t i = 10; i < space; i++)
         printf(" ");
     printf("ID: %d, Name: %s\n", root->id, root->name);
