+++ src/search_team.c
@@ -1,6 +1,6 @@
 #include "../main.h"
 
-Team* search_team_by_id(Team* node, int id) {
+Team* search_team_by_id(Team* node, size_t id) {
     if (node == NULL || node->id == id) {
         return node;
     }
@@ -31,7 +31,7 @@ Team* search_team_by_name(Team* node, const char* name) {
 
 Team* search_team(Team* root, const char* query) {
     char* nb;
-    int id = strtol(query, &nb, 10);
+    size_t id = strtol(query, &nb, 10);
 
     if (*nb == '\0') {
         return search_team_by_id(root, id);
